+++ client/components/homePage/HomeLogo.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import imagesInventory from '../../utilities/images'
+
+const HomeLogo = () => {
+  return (
+    <div id="home-logo-container">
+      <img id="home-logo" src={imagesInventory.logo} alt="Website logo" />
+    </div>
+  )
+}
+
+export default HomeLogo
