+++ client/components/navbar/material-styles.js
@@ -1,0 +1,9 @@
+//Styling for Material UI elements
+
+const styles = () => ({
+  root: {
+    flexGrow: 1
+  }
+})
+
+export default styles
