+++ client/components/searchResultsView/ResultsContainer.js
@@ -1,24 +1,20 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import {Link, withRouter} from 'react-router-dom'
-import axios from 'axios'
-import {fetchBooks} from '../../store/books.js'
 import SearchResults from './SearchResults'
 import SortDropDown from './SortDropDown'
 import {handleFilter, handleSort} from '../../utilities/sortFilterHelper'
-import styles from './styleSearchResults'
+import {resultStyles} from './material-styles'
 import {withStyles} from '@material-ui/core/styles'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import FilterMenu from './FilterMenu'
 
-//This component serves as the parent container for the search results view
+//This component displays book results base on user search input
 
 class ResultsContainer extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
-      allBooks: [], //stores the current population of books
+      allBooks: [],
       currentBooks: [], // current state of books including after user filters and/or sorts population
       sortBy: 'relevance',
       filterBy: 'everything'
@@ -34,8 +30,8 @@ class ResultsContainer extends React.Component {
     })
   }
 
-  //Update state to trigger if there are changes to the redux store when user performs a new search in navbar
   componentDidUpdate(prevProps, prevState) {
+    //Set state if there are changes to the redux store when user performs a new search in navbar
     if (prevProps.books !== prevState.allBooks) {
       this.setState({
         allBooks: this.props.books,
@@ -43,7 +39,8 @@ class ResultsContainer extends React.Component {
       })
     }
   }
-  //  Everytime selects from a sort or filter, this method will be invoked, which in turns triggers handle filter and handle sort helper functions. This allows for sorting on top of a filtered population.
+
+  //  Everytime user selects from a sort or filter, this method will be invoked, which in turns triggers handle filter and handle sort helper functions. This allows for sorting on top of a filtered population.
   async handleSortFilter(event, selection) {
     await this.setState({[event]: selection})
     let {sortBy, filterBy} = this.state
@@ -54,11 +51,12 @@ class ResultsContainer extends React.Component {
   }
 
   render() {
-    const {classes} = this.props //Use to targeting Material UI elements for styling
+    const {classes} = this.props //for Material UI styling
     const {currentBooks} = this.state
     const handleSortFilter = this.handleSortFilter
 
     if (this.props.fetchStatus)
+      //Conditional use to show loading spinner if fetching data is still in progress
       return (
         <div id="loader-container">
           <div className="loader" />
@@ -85,10 +83,7 @@ const mapStateToProps = state => ({
   fetchStatus: state.fetchStatus
 })
 
-// const mapDispatchToProps = dispatch => ({
-//   onFetchBooks: input => dispatch(fetchBooks(input))
-// })
-
 const ConnectResultsContainer = connect(mapStateToProps)(ResultsContainer)
 
-export default withStyles(styles)(ConnectResultsContainer)
+//withStyles use to attach Material UI styling to component
+export default withStyles(resultStyles)(ConnectResultsContainer)
