+++ client/components/searchResultsView/material-styles.js
@@ -1,6 +1,5 @@
 //For styling of Material UI elements for the search results view
-
-export default theme => ({
+export const resultStyles = theme => ({
   root: {},
   layout: {
     marginTop: theme.spacing.unit * 3,
@@ -8,7 +7,7 @@ export default theme => ({
     marginLeft: theme.spacing.unit * 3,
     marginRight: theme.spacing.unit * 3,
     [theme.breakpoints.up(1100 + theme.spacing.unit * 3 * 2)]: {
-      width: 1100, //adjust width of cards
+      width: 1100, //use to adjust width of cards
       marginLeft: 'auto',
       marginRight: 'auto'
     }
@@ -20,11 +19,27 @@ export default theme => ({
     display: 'flex',
     marginTop: 10
   },
-  cardDetails: {
-    // flex: 1
-  },
   cardMedia: {
     width: 160
-    //adjust thumbnail size
+  }
+})
+
+export const filterStyles = {
+  root: {
+    flexGrow: 1,
+    marginTop: 86,
+    marginLeft: 70
+  }
+}
+
+export const sortStyles = theme => ({
+  root: {
+    display: 'flex',
+    marginTop: 96,
+    marginLeft: theme.spacing.unit * 20
+  },
+  formControl: {
+    margin: theme.spacing.unit,
+    minWidth: 120
   }
 })
