+++ client/store/index.js
@@ -4,7 +4,6 @@ import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import books from './books'
 import fetchStatus from './fetchStatus'
-
 import book from './singleBook'
 
 const reducer = combineReducers({
